wagwan/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.4.2
autopep8==1.5.4
bleach==3.3.0
bootstrap4==0.1.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
Django==4.0
django-bootstrap3==21.1
django-crispy-forms==1.13.0
django-resized==0.3.11
django-shapeshifter==18.9.23
docutils==0.16
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.10
isort==5.7.0
jdcal==1.4.1
keyring==22.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
misaka==2.1.1
openpyxl==3.0.6
packaging==20.9
Pillow==8.4.0
pipenv==2020.11.15
pkginfo==1.7.0
psycopg2==2.9.2
pycodestyle==2.6.0
pycparser==2.21
Pygments==2.8.0
pylint==2.6.0
pyparsing==2.4.7
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
rope==0.18.0
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.58.0
twine==3.3.0
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
webencodings==0.5.1
wrapt==1.12.1
